Bugfix: Removed "synchronous" as per PR review. Change: Created a router as per PR review. Change: The download endpoint is now at /clinical/template/:template Enhancement: Now use the pathname to specify the template to use, which makes Chrome default to naming the file after the template, rather than calling it "template".tsv
