+++ app.js
@@ -11,9 +11,8 @@ import {PORT, NODE_ENV, GQL_MAX_COST, SUBMISSION_TEMPLATE_PATH, CLINICAL_SERVICE
 import clinicalSchema from './schemas/Clinical';
 
 import config from './package.json';
+var clinical = require('./routes/clinical');
 import costDirectiveTypeDef from './schemas/costDirectiveTypeDef';
-import urlJoin from "url-join";
-import fetch from "node-fetch";
 
 const { version } = config;
 
@@ -62,24 +61,7 @@ const init = async () => {
     res.json(version);
   });
 
-  // Our specification download service can't use GraphQL because GraphQL specification requires the content-type
-  // that it returns be json, and we want to be able to return other content types, such as tab-separated-values,
-  // so that the user is automatically prompted to save the file from their browser.
-  const apiRoot = urlJoin(CLINICAL_SERVICE_ROOT, SUBMISSION_TEMPLATE_PATH);
-  app.get('/downloads', async (req,res) => {
-    const name = req.query.template;
-    if (name == undefined) {
-      res.json({error: "Invalid query: use parameter template to specify the desired template file."});
-    } else {
-      const data = fetch(urlJoin(apiRoot, name), {synchronous: true}).then( (r) => {
-          res.status(r.status);
-          res.headers = r.headers;
-          // copying headers doesn't copy the content-type header...
-          res.contentType(r.headers.get('content-type'));
-          res.send(r.body.read());
-      });
-    }
-  });
+  app.use("/clinical", clinical);
 
   app.listen(PORT, () =>
     console.log(`🚀 Server ready at http://localhost:${PORT}${server.graphqlPath}`),
